+++ src/trade.py
@@ -1,5 +1,6 @@
 from datetime import datetime
 from enum import Enum
+from decimal import Decimal
 from pydantic.dataclasses import dataclass
 
 from src.stock import BaseStock
@@ -13,8 +14,8 @@ class BuySell(Enum):
 @dataclass
 class Trade:
     # IMPROVEMENT javier: we can have a gui easily if i move attr to
-    quantity: int
+    quantity: Decimal
     stock: BaseStock
     buysell: BuySell
-    price: float
+    price: Decimal
     timestamp: datetime = datetime.now()
